Fix foreign classes not getting generics.
@Acconut reported in lily-apache that the `Tainted` class was not
properly getting generics bound.

This changes the internal `List` and `Hash` in Lily's core to have
generics made for them. However, those two aren't actually loaded
through true class dynaload. Because of that, the lack of or addition
of generics doesn't change them. (They're autoloaded internally
because both are important, is why).