Keyword argument support for defines and variants.
This adds a new token, `Token.Keyarg` to represent keyword arguments.
The schema has a new field, keyarg, added to arguments of both a
define and a variant. The field is either set to the keyword name plus
the starting colon, or is absent.

Both bindgen and docgen support keyword arguments. Currently, docgen
writes them with the same coloring as an argument. That might change
eventually.

One caveat to watch out for is accidentally writing a definition like
so:

`define f(x: Integer, y: Integer):Integer`

Previously, `:Integer` would be seen as `:`, and `Integer`. Now, with
keyword arguments, the return type is instead seen as a keyword
argument. As a result, the engine will refuse the above (both gen
scripts will stop).